Wireframe for block party
3 - complete
4 - comfortable

Right now, the different sections are in divs because I wanted to make
the food and beer menu float next to each other, but couldn't make them
work. Was also unsure of copyright issues with photos, or what photo
editors are the best to use for those of us without design backgrounds.